refactor: Improve IP address validation in ExtensionManagementComponent
This commit updates the ExtensionManagementComponent to improve the IP address validation before submitting the form. It adds a new state variable "ipError" to display an error message if the IP address format is invalid. The validation is performed using a regular expression to match the IP address format. This change enhances the data integrity and user experience of the extension management feature.