refactor: Update DashboardComponent to use a spinner for loading state
This commit updates the DashboardComponent to use a spinner for the loading state instead of displaying the text "Loading...". It adds a new CSS file, Spinner.css, and imports it in the DashboardComponent. The spinner is displayed in the center of the screen using flexbox. This change improves the user experience by providing a visual indicator of the loading state.