feat: Add isAdmin check to AppRouter for ExtensionManagement route
This commit adds an isAdmin check to the AppRouter component for the ExtensionManagement route. If the user is authenticated, verified, and an admin, the ExtensionManagement component will be rendered. Otherwise, the user will be redirected to the NoAccessPage. This change enhances the security of the application by ensuring that only admins have access to the ExtensionManagement functionality.